Local Bug
I have a bug in the code on the local c drive, This is a last resort test to see if it is my device that is generating the error.
Hopefully this is only my device. I have a back up of the site prior should I have to start over.